Move shared code to util.py.
